Block header verification
Add two functions which together do comprehensive block header verification:
- VerifyBlockHeaderBasic, which validates header fields that only
  require the prior state.
- VerifyBlockHeaderSafrole which validates header fields which rely on
  an updated SAFROLE state to proceed.

Right now this is tested as part of safrole simluation, but more tests
to follow in a future PR.

Also:
- Fix up simulation tests which needed to have updated extrinsic hashes
  and parent state roots.
- Skip guarantees simulation test for now, updating the hashes didn't
  help because we also need to generate new signatures for the work
report after updating the state root it should point to.
- Move safrole related STF functions out into it's own file. A separate
  package would have been better, but right now that's not possible
until we structure our packages a little differently to avoid circular
deps.
- Add NextSafroleState function which can be used outside of UpdateState
  to determine the next Safrole state. This function has been moved from
simulation and updated to also potentially handle incoming disputes.
It's used only in simulation for now.
- Offenders marker will be handled later, added a TODO.
