+++ tests/simulation/disputes_block_001.json
@@ -1,7 +1,7 @@
 {
     "header": {
         "parent": "0x476243ad7cc4fc49cb6cb362c6568e931731d8650d917007a6037cceedd62244",
-        "parent_state_root": "0xc3a8849fb14674e1e0bfacf2a3cd094bcbac688330745b5b1e5a5cba88997137",
+        "parent_state_root": "0xa00b983eb2c07ef6bd4ea92f29dad2ce1903a0e960ed279d4005cd95316cce83",
         "extrinsic_hash": "0x72190a42070f29a487a554213f258b569aea109758e26347577488fd1a8d6749",
         "slot": 1,
         "epoch_mark": null,
@@ -12,7 +12,7 @@
         ],
         "author_index": 0,
         "entropy_source": "0x471e9a13ad977962ae5c68898274c0ef255ec80c9d31b26c80c2e18cf76fbc2e3dc9de282931e780db3a282fcc7abf34056404db728c64d8073c8074bc8cb51a7b6bfc64c8f902ad8e85f83a3c85bc1b2757fb3dd7ebab33c00a065823ffdc17",
-        "seal": "0xabc77c8ba19cee69325f04258b882d6515ee600e4f5da8b44b17985e096a3b290e1e5a9b9d779dcdfbb81801f9d8bb47598263d4d3b51e03ee6d3b05c0b9cc0e772b2782ddfc91f588a7d3332da98e58311e05eef582ef25dcc4f34a2629cd18"
+        "seal": "0xabc77c8ba19cee69325f04258b882d6515ee600e4f5da8b44b17985e096a3b29658902dac0402b7bdeabc42f4fce5daef261353fda2fc9cdfe3daba6ff7f5617c38ebb6a1edde20fffcc3d4f47b63d803ed6001a520e2692c6634368d7806b1a"
     },
     "extrinsic": {
         "tickets": [],
