+++ tests/simulation/guarantee_block_001.json
@@ -1,14 +1,15 @@
 {
     "header": {
         "parent": "0x476243ad7cc4fc49cb6cb362c6568e931731d8650d917007a6037cceedd62244",
-        "parent_state_root": "0x0000000000000000000000000000000000000000000000000000000000000000",
-        "extrinsic_hash": "0x0000000000000000000000000000000000000000000000000000000000000000",
+        "parent_state_root": "0x5b0937c35709936c9570ad26ba0e0832e3d1d8925dd14fbbd964178aa372d14c",
+        "extrinsic_hash": "0x8918ef4432e9b8f801450d1be106c5e826262dd692e418b5d53c67b30254bd89",
         "slot": 1,
+        "epoch_mark": null,
         "tickets_mark": null,
         "offenders_mark": [],
         "author_index": 0,
         "entropy_source": "0x471e9a13ad977962ae5c68898274c0ef255ec80c9d31b26c80c2e18cf76fbc2e3dc9de282931e780db3a282fcc7abf34056404db728c64d8073c8074bc8cb51a7b6bfc64c8f902ad8e85f83a3c85bc1b2757fb3dd7ebab33c00a065823ffdc17",
-        "seal": "0xabc77c8ba19cee69325f04258b882d6515ee600e4f5da8b44b17985e096a3b29154732fa5359bad6939a4584169d39c58d59a5c017064ea74a2ec435685a8107cc5d10e5c8aba63400e4d224d221b9d36a430258d2a70b1b73bed538200e7f03"
+        "seal": "0xabc77c8ba19cee69325f04258b882d6515ee600e4f5da8b44b17985e096a3b29858a45ca44b6e151421d3a0f01734f8d6c82a6f20406ad225f4d205f79c82113535b59ed6690dce20aed07d3068bda9af75bcf45fee4f1622526c5f02beb4114"
     },
     "extrinsic": {
         "tickets": [],
@@ -25,7 +26,7 @@
                     },
                     "context": {
                         "anchor": "0x476243ad7cc4fc49cb6cb362c6568e931731d8650d917007a6037cceedd62244",
-                        "state_root": "0x0000000000000000000000000000000000000000000000000000000000000000",
+                        "state_root": "0x5b0937c35709936c9570ad26ba0e0832e3d1d8925dd14fbbd964178aa372d14c",
                         "beefy_root": "0x0000000000000000000000000000000000000000000000000000000000000000",
                         "lookup_anchor": "0x60751ab5b251361fbfd3ad5b0e84f051ccece6b00830aed31a5354e00b20b9ed",
                         "lookup_anchor_slot": 33,
@@ -75,15 +76,15 @@
                 "signatures": [
                     {
                         "validator_index": 1,
-                        "signature": "0xa5d972c5603d0d999419883331957726edd9b29af8efa8c0780b6e028ba1817c17cc00ccd41108bb7907c39876d5d6d5ee249c484f070644e53de48144786200"
+                        "signature": "0x81ff9abb16df2eefa5d46820a820a747ee0b8ddcef85161780a19415caf455ffe955fc4a074149064549df8df39188273f7c630fb5c6ab658455efeb852da804"
                     },
                     {
                         "validator_index": 2,
-                        "signature": "0xbb0cd2359afb661fd6e82f8a7a6da412260068cff9530ff7a1fa5eeb6f923371ecba4635640b2c524af55faf940617b94cdc67243e2b5d0fafd69690d70bd206"
+                        "signature": "0x75719399973df76a6db425a03289eb135779a579b1d2349b6c14a74032692549aec70b7b1c1c54ead1778907ea19817d7779ad82a5079c383dcf886eb2e0d200"
                     },
                     {
                         "validator_index": 5,
-                        "signature": "0x92414235e8c029bd3bb367425eff7bb319f7f9a5727ed978a378eeccb01616d95497fcf75edae16d31827776ac4cdfece660a978678f766c572719060835d305"
+                        "signature": "0x633701bed3277ac91d691517f484e75ed38a5e9fd3b1f6ad4ea15bf49aff17bc781a8a44454cf75dc9ab1741f51921f86357701138266ab79f26a39de40f1707"
                     }
                 ]
             }
