+++ tests/simulation/preimage_block_001.json
@@ -1,7 +1,7 @@
 {
     "header": {
         "parent": "0x476243ad7cc4fc49cb6cb362c6568e931731d8650d917007a6037cceedd62244",
-        "parent_state_root": "0xdc7695047ae21f0e8dd14b94c8a96ed82c7fcefaf064da03876e804673a6ddb2",
+        "parent_state_root": "0x41787e62f2dc023436a018f1de429286d817f79d5c87f853191bf6bce2a04796",
         "extrinsic_hash": "0x9b567bdcf5a2c78fd328ab706f8c32f319f95e0ae28126185c6b10d2a38b3189",
         "slot": 1,
         "epoch_mark": null,
@@ -9,7 +9,7 @@
         "offenders_mark": [],
         "author_index": 0,
         "entropy_source": "0x471e9a13ad977962ae5c68898274c0ef255ec80c9d31b26c80c2e18cf76fbc2e3dc9de282931e780db3a282fcc7abf34056404db728c64d8073c8074bc8cb51a7b6bfc64c8f902ad8e85f83a3c85bc1b2757fb3dd7ebab33c00a065823ffdc17",
-        "seal": "0xabc77c8ba19cee69325f04258b882d6515ee600e4f5da8b44b17985e096a3b29154732fa5359bad6939a4584169d39c58d59a5c017064ea74a2ec435685a8107cc5d10e5c8aba63400e4d224d221b9d36a430258d2a70b1b73bed538200e7f03"
+        "seal": "0xabc77c8ba19cee69325f04258b882d6515ee600e4f5da8b44b17985e096a3b29db6ac0b709511cbc47023e64133638df1d587fd516889d23e0519bea13117a02c117d5c3ef23c5f1986deba94da6d1b832c15c9dc73454b04c84514137b9fb0a"
     },
     "extrinsic": {
         "tickets": [],
