+++ tests/simulation/simulation_assurance_test.go
@@ -57,6 +57,9 @@ func TestSimulateAssurance(t *testing.T) {
 		require.NoError(t, err, "failed to close db")
 	}()
 
+	trieDB := store.NewTrie(db)
+	require.NoError(t, err)
+
 	chainDB := store.NewChain(db)
 	require.NoError(t, err)
 
@@ -70,6 +73,7 @@ func TestSimulateAssurance(t *testing.T) {
 		currentState,
 		testBlock,
 		chainDB,
+		trieDB,
 	)
 	require.NoError(t, err)
